rename package name from partitions to partitioning in common module
