+++ common/src/main/java/com/datastrato/gravitino/dto/rel/partitioning/HourPartitioningDTO.java
@@ -2,20 +2,17 @@
  * Copyright 2023 Datastrato Pvt Ltd.
  * This software is licensed under the Apache License version 2.
  */
-package com.datastrato.gravitino.dto.rel.partitions;
+package com.datastrato.gravitino.dto.rel.partitioning;
 
 import com.google.common.base.Preconditions;
-import lombok.EqualsAndHashCode;
 import org.apache.commons.lang3.ArrayUtils;
 
-@EqualsAndHashCode(callSuper = true)
-public final class DayPartitioningDTO extends Partitioning.SingleFieldPartitioning {
-
-  public static DayPartitioningDTO of(String... fieldName) {
-    return new DayPartitioningDTO(fieldName);
+public final class HourPartitioningDTO extends Partitioning.SingleFieldPartitioning {
+  public static HourPartitioningDTO of(String[] fieldName) {
+    return new HourPartitioningDTO(fieldName);
   }
 
-  private DayPartitioningDTO(String[] fieldName) {
+  private HourPartitioningDTO(String[] fieldName) {
     Preconditions.checkArgument(
         ArrayUtils.isNotEmpty(fieldName), "fieldName cannot be null or empty");
     this.fieldName = fieldName;
@@ -23,6 +20,6 @@ public final class DayPartitioningDTO extends Partitioning.SingleFieldPartitioni
 
   @Override
   public Strategy strategy() {
-    return Strategy.DAY;
+    return Strategy.HOUR;
   }
 }
