+++ common/src/main/java/com/datastrato/gravitino/dto/rel/partitioning/IdentityPartitioningDTO.java
@@ -2,7 +2,7 @@
  * Copyright 2023 Datastrato Pvt Ltd.
  * This software is licensed under the Apache License version 2.
  */
-package com.datastrato.gravitino.dto.rel.partitions;
+package com.datastrato.gravitino.dto.rel.partitioning;
 
 import com.google.common.base.Preconditions;
 import lombok.EqualsAndHashCode;
