+++ common/src/main/java/com/datastrato/gravitino/dto/rel/partitioning/MonthPartitioningDTO.java
@@ -2,17 +2,17 @@
  * Copyright 2023 Datastrato Pvt Ltd.
  * This software is licensed under the Apache License version 2.
  */
-package com.datastrato.gravitino.dto.rel.partitions;
+package com.datastrato.gravitino.dto.rel.partitioning;
 
 import com.google.common.base.Preconditions;
 import org.apache.commons.lang3.ArrayUtils;
 
-public final class YearPartitioningDTO extends Partitioning.SingleFieldPartitioning {
-  public static YearPartitioningDTO of(String[] fieldName) {
-    return new YearPartitioningDTO(fieldName);
+public final class MonthPartitioningDTO extends Partitioning.SingleFieldPartitioning {
+  public static MonthPartitioningDTO of(String[] fieldName) {
+    return new MonthPartitioningDTO(fieldName);
   }
 
-  private YearPartitioningDTO(String[] fieldName) {
+  private MonthPartitioningDTO(String[] fieldName) {
     Preconditions.checkArgument(
         ArrayUtils.isNotEmpty(fieldName), "fieldName cannot be null or empty");
     this.fieldName = fieldName;
@@ -20,6 +20,6 @@ public final class YearPartitioningDTO extends Partitioning.SingleFieldPartition
 
   @Override
   public Strategy strategy() {
-    return Strategy.YEAR;
+    return Strategy.MONTH;
   }
 }
