+++ common/src/main/java/com/datastrato/gravitino/dto/rel/partitioning/YearPartitioningDTO.java
@@ -2,17 +2,17 @@
  * Copyright 2023 Datastrato Pvt Ltd.
  * This software is licensed under the Apache License version 2.
  */
-package com.datastrato.gravitino.dto.rel.partitions;
+package com.datastrato.gravitino.dto.rel.partitioning;
 
 import com.google.common.base.Preconditions;
 import org.apache.commons.lang3.ArrayUtils;
 
-public final class MonthPartitioningDTO extends Partitioning.SingleFieldPartitioning {
-  public static MonthPartitioningDTO of(String[] fieldName) {
-    return new MonthPartitioningDTO(fieldName);
+public final class YearPartitioningDTO extends Partitioning.SingleFieldPartitioning {
+  public static YearPartitioningDTO of(String[] fieldName) {
+    return new YearPartitioningDTO(fieldName);
   }
 
-  private MonthPartitioningDTO(String[] fieldName) {
+  private YearPartitioningDTO(String[] fieldName) {
     Preconditions.checkArgument(
         ArrayUtils.isNotEmpty(fieldName), "fieldName cannot be null or empty");
     this.fieldName = fieldName;
@@ -20,6 +20,6 @@ public final class MonthPartitioningDTO extends Partitioning.SingleFieldPartitio
 
   @Override
   public Strategy strategy() {
-    return Strategy.MONTH;
+    return Strategy.YEAR;
   }
 }
