+++ trino-connector/src/main/java/com/datastrato/gravitino/trino/connector/catalog/iceberg/IcebergMetadataAdapter.java
@@ -8,7 +8,7 @@ import static com.datastrato.gravitino.trino.connector.catalog.iceberg.IcebergPr
 import static com.datastrato.gravitino.trino.connector.catalog.iceberg.IcebergPropertyMeta.ICEBERG_SORTED_BY_PROPERTY;
 
 import com.datastrato.catalog.property.PropertyConverter;
-import com.datastrato.gravitino.dto.rel.partitions.Partitioning.SingleFieldPartitioning;
+import com.datastrato.gravitino.dto.rel.partitioning.Partitioning;
 import com.datastrato.gravitino.rel.expressions.Expression;
 import com.datastrato.gravitino.rel.expressions.NamedReference;
 import com.datastrato.gravitino.rel.expressions.sorts.SortOrder;
@@ -150,7 +150,7 @@ public class IcebergMetadataAdapter extends CatalogConnectorMetadataAdapter {
           ICEBERG_PARTITIONING_PROPERTY,
           gravitinoTable.getPartitioning().length > 0
               ? Arrays.stream(gravitinoTable.getPartitioning())
-                  .map(ts -> ((SingleFieldPartitioning) ts).fieldName()[0])
+                  .map(ts -> ((Partitioning.SingleFieldPartitioning) ts).fieldName()[0])
                   .collect(Collectors.toList())
               : Collections.EMPTY_LIST);
     }
